:hammer: Refactoriza codigo en modal.jsx tanto mobile como desktop, ajusta boton register.jsx
